refactor for better testability and code organization
